Move slice manipulation to the unsafe package
