+++ internal/unsafe/slice_conversion.go
@@ -19,24 +19,24 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-package uuid
+package unsafe
 
 import (
 	"unsafe"
 )
 
-// byteSliceToString uses pointer manipulation tricks to take a slice of bytes
+// ByteSliceToString uses pointer manipulation tricks to take a slice of bytes
 // and convert it into a string. It should be used to avoid copying the slice.
 // It should also cause the compiler's escape analysis to keep the slice on
 // the stack if it would otherwise be stack allocated.
-func byteSliceToString(b []byte) string {
+func ByteSliceToString(b []byte) string {
 	return *(*string)(unsafe.Pointer(&b))
 }
 
-// stringToByteSlice uses pointer manipulation tricks to take a string and
+// StringToByteSlice uses pointer manipulation tricks to take a string and
 // return the underlying slice of bytes referenced by the string. This is
 // useful when wanting to avoid extra allocations.
-func stringToByteSlice(s string) []byte {
+func StringToByteSlice(s string) []byte {
 	return *(*[]byte)(
 		unsafe.Pointer(
 			&struct {
